+++ src/fairshift_lab/__init__.py
@@ -3,12 +3,15 @@
 from fairshift_lab.calibration import apply_temperature, calibration_summary, fit_temperature
 from fairshift_lab.config import ExperimentConfig, ShiftConfig, ShiftKind
 from fairshift_lab.experiment import ExperimentResult, run_experiment
+from fairshift_lab.policy_study import PolicyStudyConfig, PolicyStudyResult, run_policy_study
 from fairshift_lab.sensitivity import threshold_sweep
 from fairshift_lab.study import StudyConfig, StudyResult, run_study
 
 __all__ = [
     "ExperimentConfig",
     "ExperimentResult",
+    "PolicyStudyConfig",
+    "PolicyStudyResult",
     "ShiftConfig",
     "ShiftKind",
     "StudyConfig",
@@ -17,7 +20,8 @@ __all__ = [
     "calibration_summary",
     "fit_temperature",
     "run_experiment",
+    "run_policy_study",
     "run_study",
     "threshold_sweep",
 ]
-__version__ = "1.0.1"
+__version__ = "1.1.0"